% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-coxphMIC.R
\name{plot.coxphMIC}
\alias{plot.coxphMIC}
\title{The Generic \code{plot} Function for Object of \code{coxphMIC} Class}
\usage{
\method{plot}{coxphMIC}(x, conf.level = 0.95, horizontal = TRUE,
  mar = rep(4.5, 4), ...)
}
\arguments{
\item{x}{an object of \code{coxphMIC} class.}

\item{conf.level}{confidence level used for error bar plots. Default is 0.95.}

\item{horizontal}{Logical indicator of horizontal alignment. Default is \code{TRUE}.}

\item{mar}{margin in terms of the number of lines to be specified on the four sides of the plot}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Error bar plots for estimated gamma and beta at a given confidenve level.
}
\description{
The Generic \code{plot} Function for Object of \code{coxphMIC} Class
}
\details{
The (generic) plot method for an \code{coxphMIC} object. It plots MIC estimates of
gamma and beta. For 0 beta estimates, their corresponding
SE are reset to 0 to make the plot.
}
\references{
\itemize{
\item Abdolyousefi, R. N. and Su, X. (2016). \bold{coxphMIC}: An R package for sparse estimation of Cox PH Models via approximated information criterion. Tentatively accepted, \emph{The R Journal}.
\item Su, X. (2015). Variable selection via subtle uprooting.
\emph{Journal of Computational and Graphical Statistics}, \bold{24}(4): 1092--1113.
URL \url{http://www.tandfonline.com/doi/pdf/10.1080/10618600.2014.955176}
\item Su, X., Wijayasinghe, C. S., Fan, J., and Zhang, Y. (2015). Sparse estimation of Cox proportional
hazards models via approximated information criteria. \emph{Biometrics}, \bold{72}(3): 751--759.
URL \url{http://onlinelibrary.wiley.com/doi/10.1111/biom.12484/epdf}
}
}
\seealso{
\code{\link[coxphMIC]{coxphMIC}}
}
